+++ qfi_opt/calculate_qfi_example.py
@@ -3,7 +3,7 @@ import itertools
 
 import numpy as np
 
-import spin_models
+from qfi_opt import spin_models
 
 
 def variance(rho: np.ndarray, G: np.ndarray) -> float:
