+++ qfi_opt/dissipation.py
