+++ qfi_opt/spin_models.py
