+++ requirements.txt
@@ -2,5 +2,4 @@ jax>=0.4.8
 jaxlib>=0.4.7
 matplotlib>=3.6.2
 numpy>=1.23.4
-pytest>=7.3.0
 scipy>=1.9.2
