+++ setup.cfg
@@ -1,9 +1,5 @@
 # Configurations for our checks
 
-[tool:black]
-line-length = 160
-target-version = py310
-
 [isort]
 profile = black
 color_output = True
